requirements.txt
func_timeout
scipy
netCDF4
requests
python-dotenv
